sql/planner/optimizer/mod.rs -> sql/planner/optimizer/optimizer.rs
